pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wsws-tts"
version = "0.1.0"
description = "CLI: extract WSWS.org articles and convert to speech via Resemble AI Chatterbox"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31",
  "beautifulsoup4>=4.12",
  "lxml>=5.1",
  "numpy>=1.26",
  "chatterbox-tts>=0.1.0",
]

[project.scripts]
wsws-tts = "wsws_tts.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
